force int for ENAUG

+++ custodian/vasp/handlers.py
@@ -642,7 +642,7 @@ class DriftErrorHandler(ErrorHandler):
         # PREC is already high and ENAUG set so just increase it
         else:
             actions.append({"dict": "INCAR",
-                            "action": {"_set": {"ENAUG": incar.get("ENAUG", 2080) * self.enaug_multiply}}})
+                            "action": {"_set": {"ENAUG": int(incar.get("ENAUG", 2080) * self.enaug_multiply)}}})
 
         curr_drift = np.sum(outcar.data.get('drift')[-1 * self.to_average:]) / (3 * self.to_average)
         VaspModder(vi=vi).apply_actions(actions)
